new container classes

+++ LCUD_Project/Cube.h
@@ -1,13 +1,12 @@
 #ifndef CUBE_H
 #define CUBE_H
-#include <string>
 #include "Container.cpp"
 
      
 class Cube : public Container{
     private:
     public:
-        Cube(float side);
+        Cube(int side);
         ~Cube(){;}
         std::string to_string();
